Integrate information need in US
