+++ tests/conftest.py
@@ -4,7 +4,9 @@ import os
 
 import pytest
 
+from usersimcrs.core.information_need import InformationNeed
 from usersimcrs.core.simulation_domain import SimulationDomain
+from usersimcrs.items.item import Item
 from usersimcrs.items.item_collection import ItemCollection
 
 DOMAIN_YAML_FILE = "tests/data/domains/movies.yaml"
@@ -42,3 +44,22 @@ def item_collection(domain: SimulationDomain):
     )
     yield item_collection
     os.remove("tests/data/items.db")
+
+
+@pytest.fixture
+def information_need() -> InformationNeed:
+    """Information need fixture."""
+    constraints = {"GENRE": "Comedy", "DIRECTOR": "Steven Spielberg"}
+    requests = ["PLOT", "RATING"]
+    target_items = [
+        Item(
+            "1",
+            {
+                "GENRE": "Comedy",
+                "DIRECTOR": "Steven Spielberg",
+                "RATING": 4.5,
+                "PLOT": "A movie plot",
+            },
+        )
+    ]
+    return InformationNeed(target_items, constraints, requests)
