+++ usersimcrs/simulator/user_simulator.py
@@ -5,14 +5,29 @@ from abc import ABC, abstractmethod
 from dialoguekit.core.utterance import Utterance
 from dialoguekit.participant.user import User, UserType
 
+from usersimcrs.core.information_need import generate_random_information_need
+from usersimcrs.core.simulation_domain import SimulationDomain
+from usersimcrs.items.item_collection import ItemCollection
+
 
 class UserSimulator(User, ABC):
     def __init__(
         self,
         id: str,
+        domain: SimulationDomain,
+        item_collection: ItemCollection,
     ) -> None:
         """Initializes the user simulator."""
         super().__init__(id, UserType.SIMULATOR)
+        self._domain = domain
+        self._item_collection = item_collection
+        self.get_new_information_need()
+
+    def get_new_information_need(self) -> None:
+        """Generates a new information need."""
+        self.information_need = generate_random_information_need(
+            self._domain, self._item_collection
+        )
 
     @abstractmethod
     def _generate_response(self, agent_utterance: Utterance) -> Utterance:
